apps/order-service/requirements.txt
fastapi
uvicorn[standard]
httpx
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
